refactor(kimi-code): move the file-change warning copy into the TUI constant directory
