+++ apps/kimi-code/src/tui/constant/kimi-tui.ts
@@ -4,6 +4,8 @@ export { DEFAULT_OAUTH_PROVIDER_NAME, OAUTH_LOGIN_REQUIRED_CODE, PRODUCT_NAME } 
 
 export const LLM_NOT_SET_MESSAGE = 'LLM not set, send "/login" to login';
 export const NO_ACTIVE_SESSION_MESSAGE = 'No active session. Send /login to login.';
+export const UNCONFIRMED_FILE_CHANGES_WARNING =
+  'In this mode, Kimi Code can modify or delete files without your confirmation';
 export const CTRL_D_HINT = 'Press Ctrl+D again to exit';
 export const CTRL_C_HINT = 'Press Ctrl+C again to exit';
 export const MAIN_AGENT_ID = 'main';
